+++ gen-graph/src/lib.rs
@@ -681,150 +681,6 @@ mod tests {
     }
 
     #[test]
-    fn merges_gen_graphs_and_preserves_distinct_edges() {
-        let start = GraphNode {
-            node_id: HashId::convert_str("merge-start"),
-            sequence_start: 0,
-            sequence_end: 0,
-        };
-        let middle = GraphNode {
-            node_id: HashId::convert_str("merge-middle"),
-            sequence_start: 0,
-            sequence_end: 5,
-        };
-        let end = GraphNode {
-            node_id: HashId::convert_str("merge-end"),
-            sequence_start: 0,
-            sequence_end: 0,
-        };
-        let edge_a = GraphEdge {
-            edge_id: HashId::convert_str("merge-edge-a"),
-            source_strand: Strand::Forward,
-            target_strand: Strand::Forward,
-            chromosome_index: 0,
-            phased: 0,
-            created_on: 1,
-        };
-        let edge_b = GraphEdge {
-            edge_id: HashId::convert_str("merge-edge-b"),
-            source_strand: Strand::Forward,
-            target_strand: Strand::Forward,
-            chromosome_index: 1,
-            phased: 1,
-            created_on: 2,
-        };
-
-        let mut graph_a = GenGraph::new();
-        graph_a.add_edge(start, middle, vec![edge_a]);
-
-        let mut graph_b = GenGraph::new();
-        graph_b.add_edge(start, middle, vec![edge_b]);
-        graph_b.add_edge(middle, end, vec![edge_a]);
-
-        graph_a.merge_graph(&graph_b);
-
-        assert_eq!(graph_a.edge_weight(start, middle).unwrap().len(), 2);
-        assert!(
-            graph_a
-                .edge_weight(start, middle)
-                .unwrap()
-                .contains(&edge_a)
-        );
-        assert!(
-            graph_a
-                .edge_weight(start, middle)
-                .unwrap()
-                .contains(&edge_b)
-        );
-        assert_eq!(graph_a.edge_weight(middle, end).unwrap(), &vec![edge_a]);
-    }
-
-    #[test]
-    fn merges_gen_graphs_without_duplicating_identical_edges() {
-        let start = GraphNode {
-            node_id: HashId::convert_str("dedup-start"),
-            sequence_start: 0,
-            sequence_end: 0,
-        };
-        let end = GraphNode {
-            node_id: HashId::convert_str("dedup-end"),
-            sequence_start: 0,
-            sequence_end: 0,
-        };
-        let edge = GraphEdge {
-            edge_id: HashId::convert_str("dedup-edge"),
-            source_strand: Strand::Forward,
-            target_strand: Strand::Forward,
-            chromosome_index: 0,
-            phased: 0,
-            created_on: 1,
-        };
-
-        let mut graph_a = GenGraph::new();
-        graph_a.add_edge(start, end, vec![edge]);
-
-        let mut graph_b = GenGraph::new();
-        graph_b.add_edge(start, end, vec![edge]);
-
-        graph_a.merge_graph(&graph_b);
-
-        assert_eq!(graph_a.edge_weight(start, end).unwrap(), &vec![edge]);
-    }
-
-    #[test]
-    fn merges_gen_graphs_by_graph_node_value() {
-        let source_a = GraphNode {
-            node_id: HashId::convert_str("shared-node"),
-            sequence_start: 0,
-            sequence_end: 5,
-        };
-        let target_a = GraphNode {
-            node_id: HashId::convert_str("target-node"),
-            sequence_start: 0,
-            sequence_end: 5,
-        };
-        let source_b = GraphNode {
-            node_id: source_a.node_id,
-            sequence_start: source_a.sequence_start,
-            sequence_end: source_a.sequence_end,
-        };
-        let target_b = GraphNode {
-            node_id: target_a.node_id,
-            sequence_start: target_a.sequence_start,
-            sequence_end: target_a.sequence_end,
-        };
-        let edge_a = GraphEdge {
-            edge_id: HashId::convert_str("logical-edge-a"),
-            source_strand: Strand::Forward,
-            target_strand: Strand::Forward,
-            chromosome_index: 0,
-            phased: 0,
-            created_on: 1,
-        };
-        let edge_b = GraphEdge {
-            edge_id: HashId::convert_str("logical-edge-b"),
-            source_strand: Strand::Forward,
-            target_strand: Strand::Forward,
-            chromosome_index: 1,
-            phased: 1,
-            created_on: 2,
-        };
-
-        let mut graph_a = GenGraph::new();
-        graph_a.add_edge(source_a, target_a, vec![edge_a]);
-
-        let mut graph_b = GenGraph::new();
-        graph_b.add_edge(source_b, target_b, vec![edge_b]);
-
-        graph_a.merge_graph(&graph_b);
-
-        assert_eq!(graph_a.nodes().count(), 2);
-        assert!(graph_a.nodes().any(|node| node == source_a));
-        assert!(graph_a.nodes().any(|node| node == target_a));
-        assert_eq!(graph_a.edge_weight(source_a, target_a).unwrap().len(), 2);
-    }
-
-    #[test]
     fn test_three_by_three_combinatorial_graph() {
         let mut graph: DiGraphMap<i64, ()> = DiGraphMap::new();
         graph.add_node(1);
